Novas rotas de categorias e noticias

+++ pages/api/categoria.ts
@@ -1,26 +1,24 @@
 // routes/categoria.js
-import { NextApiRequest, NextApiResponse } from 'next';
-import { CategoriaModel } from '../../Models/CategoriaModel';
-import nc from 'next-connect';
-import { conectarMongoDB } from '../../middlewares/conectarMongoDB';
-import { validarTokenJwt } from '../../middlewares/validarTokenJWT';
-import { politicaCORS } from '../../middlewares/politicaCORS';
-import { RespostaPadraoMsg } from '../../types/RespostaPadraoMsg';
+import { NextApiRequest, NextApiResponse } from "next";
+import { CategoriaModel } from "../../Models/CategoriaModel";
+import nc from "next-connect";
+import { conectarMongoDB } from "../../middlewares/conectarMongoDB";
+import { validarTokenJwt } from "../../middlewares/validarTokenJWT";
+import { politicaCORS } from "../../middlewares/politicaCORS";
+import { RespostaPadraoMsg } from "../../types/RespostaPadraoMsg";
 
-
-
-const handler = nc()
-
-  .post(async (req: any, res: NextApiResponse<RespostaPadraoMsg | any>) => {
+const handler = nc().post(
+  async (req: any, res: NextApiResponse<RespostaPadraoMsg | any>) => {
     try {
       let { nomeCategoria } = req.body;
-      nomeCategoria = nomeCategoria.toLowerCase();
-
+      // nomeCategoria = nomeCategoria.toLowerCase();
 
       // Verifique se o nome da categoria já existe
-      const categoriaExistente = await CategoriaModel.findOne({ nomeCategoria });
+      const categoriaExistente = await CategoriaModel.findOne({
+        nomeCategoria,
+      });
       if (categoriaExistente) {
-        return res.status(400).json({ erro: 'Categoria já existe' });
+        return res.status(400).json({ erro: "Categoria já existe" });
       }
 
       // Crie uma nova categoria
@@ -30,70 +28,75 @@ const handler = nc()
       return res.status(201).json(novaCategoria);
     } catch (error) {
       console.error(error);
-      return res.status(500).json({ erro: 'Erro ao criar categoria' });
-
-    };
-  });
+      return res.status(500).json({ erro: "Erro ao criar categoria" });
+    }
+  }
+);
 handler.put(async (req: any, res: NextApiResponse<RespostaPadraoMsg | any>) => {
   try {
     let { nomeCategoria, id } = req.body;
-    nomeCategoria = nomeCategoria.toLowerCase();
 
     // Encontre a categoria pelo ID e atualize o nome
-    const categoriaAtualizada = await CategoriaModel.findByIdAndUpdate(id, { nomeCategoria }, { new: true });
+    const categoria = await CategoriaModel.findById(id);
 
-    if (!categoriaAtualizada) {
-      return res.status(400).json({ erro: 'Categoria não encontrada' });
+    if (!categoria) {
+      return res.status(400).json({ erro: "Categoria não encontrada" });
     }
+
+    // Encontre a categoria pelo ID e atualize o nome
+    const categoriaAtualizada = [id, nomeCategoria];
+
+    await CategoriaModel.findByIdAndUpdate(id, categoriaAtualizada);
 
     return res.status(200).json(categoriaAtualizada);
   } catch (error) {
     console.error(error);
-    return res.status(500).json({ erro: 'Erro ao atualizar categoria' });
+    return res.status(500).json({ erro: "Erro ao atualizar categoria" });
   }
 });
 
-handler.delete(async (req: any, res: NextApiResponse<RespostaPadraoMsg | any>) => {
-  try {
-    const { id } = req.body;
+handler.delete(
+  async (req: any, res: NextApiResponse<RespostaPadraoMsg | any>) => {
+    try {
+      const { id } = req.query;
 
-    // Encontre a categoria pelo ID e delete
-    const categoriaDeletada = await CategoriaModel.findByIdAndDelete(id);
+      // Encontre a categoria pelo ID e delete
+      const categoria = await CategoriaModel.findById(id);
 
-    if (!categoriaDeletada) {
-      return res.status(400).json({ erro: 'Categoria não encontrada' });
+      if (!categoria) {
+        return res.status(400).json({ erro: "Categoria não encontrada" });
+      }
+
+      await CategoriaModel.findByIdAndDelete(id);
+
+      return res
+        .status(200)
+        .json({ mensagem: "Categoria deletada com sucesso" });
+    } catch (error) {
+      console.error(error);
+      return res.status(500).json({ erro: "Erro ao deletar categoria" });
     }
-
-    return res.status(200).json({ mensagem: 'Categoria deletada com sucesso' });
-  } catch (error) {
-    console.error(error);
-    return res.status(500).json({ erro: 'Erro ao deletar categoria' });
   }
-});
-
+);
 
 handler.get(async (req: any, res: NextApiResponse<RespostaPadraoMsg | any>) => {
   try {
-    if (req.method !== 'GET') {
-      return res.status(405).json({ erro: 'Método não permitido' });
-
+    if (req.method !== "GET") {
+      return res.status(405).json({ erro: "Método não permitido" });
     }
 
-    const categorias = await CategoriaModel.find();
+    const categorias = await CategoriaModel.find().sort({ nomeCategoria: 1 });
 
     const categoriasFormatadas = categorias.map((categoria) => ({
       nomeCategoria: categoria.nomeCategoria,
       _id: categoria._id.toString(),
     }));
 
-
-
     return res.status(200).json({ categorias: categoriasFormatadas });
   } catch (error) {
     console.error(error);
-    return res.status(500).json({ erro: 'Erro ao listar categorias' });
+    return res.status(500).json({ erro: "Erro ao listar categorias" });
   }
 });
-
 
 export default politicaCORS(validarTokenJwt(conectarMongoDB(handler)));
